requirements.txt

# Core voice processing
vosk==0.3.45
pyttsx3==2.90
pyaudio==0.2.11
speechrecognition==3.10.0

# AI and NLP
ollama==0.1.7
spacy==3.7.2
chromadb==0.4.15
sentence-transformers==2.2.2

# Audio visualization
pygame==2.5.2
numpy==1.24.3
matplotlib==3.7.2

# System and utilities
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
rich==13.5.2

# Development
pytest==7.4.2
black==23.7.0
flake8==6.0.0
